Add yaf and Update versions.txt

+++ include/memory.sh
@@ -38,7 +38,7 @@ elif [ $Mem -gt 8000 ]; then
 fi
 
 # add swapfile
-if [ "${Swap}" == '0' ] && [ ${Mem} -le 2048 ]; then
+if [ ! -e ~/.oneinstack ] && [ "${Swap}" == '0' ] && [ ${Mem} -le 2048 ]; then
   echo "${CWARNING}Add Swap file, It may take a few minutes... ${CEND}"
   dd if=/dev/zero of=/swapfile count=2048 bs=1M
   mkswap /swapfile
